Moved RuleVariable into standalone package.
