+++ core-rules/src/main/java/org/hisp/dhis/android/rules/ValueMapFactory.java
@@ -1,7 +1,7 @@
 package org.hisp.dhis.android.rules;
 
 import org.hisp.dhis.android.rules.models.Event;
-import org.hisp.dhis.android.rules.models.RuleVariable;
+import org.hisp.dhis.android.rules.models.variables.RuleVariable;
 import org.hisp.dhis.android.rules.models.TrackedEntityAttributeValue;
 import org.hisp.dhis.android.rules.models.TrackedEntityDataValue;
 import org.hisp.dhis.android.rules.models.ValueType;
